Move Globalstyles to components directory
