+++ src/index.js
@@ -2,13 +2,12 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
 import * as serviceWorker from './serviceWorker'
-import GlobalStyles from './GlobalStyles'
+import GlobalStyles from './components/GlobalStyles'
 
 ReactDOM.render(
   <React.StrictMode>
-    <GlobalStyles>
-      <App />
-    </GlobalStyles>
+    <GlobalStyles />
+    <App />
   </React.StrictMode>,
   document.getElementById('root')
 )
